refactor: cdxgen structure
